feat(api): integrate Spatie Media Library and update Filament resources
Spatie Media Library Integration:
- Configure Media model with Spatie MediaLibrary
- Set up MinIO/S3 storage for media files
- Add media conversions and collections
- Update Filament panels to use Spatie media uploader

Filament Updates:
- Update all admin resources for better UX
- Add PageResource for CMS page management
- Configure custom login/logout responses
- Update panel providers with navigation groups
- Improve widget layouts and statistics

CMS Configuration:
- Configure filament-flexible-content-block-pages
- Set up content blocks for dynamic pages
- Add support for multilingual content

Media Storage:
- Configure conversions: thumb (150x150), medium (300x300)
- Set up responsive images support
- Enable MinIO/S3 for production-ready storage

Files Updated:
- All Filament Admin Resources
- AppServiceProvider, AdminPanelProvider, VendorPanelProvider
- Media model with HasMedia trait
- ListingController for media handling

Refs: Media Storage Setup, Filament Admin Enhancement

+++ apps/laravel-api/app/Filament/Admin/Widgets/PlatformStatsWidget.php
@@ -16,10 +16,10 @@ class PlatformStatsWidget extends BaseWidget
 {
     protected function getStats(): array
     {
-        $totalRevenue = Booking::where('status', BookingStatus::CONFIRMED)
+        $totalRevenue = (float) Booking::where('status', BookingStatus::CONFIRMED)
             ->sum('total_amount');
 
-        $monthlyRevenue = Booking::where('status', BookingStatus::CONFIRMED)
+        $monthlyRevenue = (float) Booking::where('status', BookingStatus::CONFIRMED)
             ->whereMonth('created_at', now()->month)
             ->sum('total_amount');
